+++ packages/workspace-utils/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @prairielearn/workspace-utils
 
+## 1.0.23
+
+### Patch Changes
+
+- Updated dependencies [fd8f6e6]
+  - @prairielearn/postgres@2.0.2
+
+## 1.0.22
+
+### Patch Changes
+
+- Updated dependencies [901fce8]
+  - @prairielearn/postgres@2.0.1
+
 ## 1.0.21
 
 ### Patch Changes